requirements.txt
absl-py==0.12.0
albumentations==1.0.0
antlr4-python3-runtime==4.8
av==8.0.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.2
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
docker-pycreds==0.4.0
filelock==3.0.12
future==0.18.2
fvcore
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
iopath==0.1.8
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
networkx==2.5.1
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.0.dev22
opencv-python-headless==4.5.2.52
packaging==20.9
parameterized==0.8.1
pathtools==0.1.2
Pillow==8.2.0
pip==21.0.1
portalocker==1.7.0
promise==2.3
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytorchvideo==0.1.0
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.1
scipy==1.6.3
sentry-sdk==1.1.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.4.8
timm==0.4.5
tokenizers==0.10.2
torch==1.8.0
torchvision==0.9.0
tqdm==4.60.0
transformers==4.5.1
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.31
Werkzeug==1.0.1
wheel==0.36.2
yacs==0.1.6
zipp==3.4.1
easydict
einops
